outlined a get_best_window function
the idea is that you pass in one of the dictionaries of a given metric,
and it gives you back the best window position for that metric. The
(smallish) challenge is what to do if there’s >1 ‘best’ window
position. My suggestion is to take the first longest one (i.e. if
there’s more than one longest one, take the first).